+++ Assets/Script/HeroStateMachine/Damage/HeroTakeDamage.cs
@@ -10,7 +10,7 @@ public class HeroTakeDamage : TakeDamageController
     {
         this.impulseSource = this.transform.GetComponent<CinemachineImpulseSource>();
     }
-    public override void TakeDamage(GameObject target, float getDamageAmount)
+    public override void TakeDamage(GameObject target, int getDamageAmount)
     {
         HeroStateMachine hsm = target.GetComponent<HeroStateMachine>();
         this.DamagePop(hsm, getDamageAmount);
@@ -18,14 +18,14 @@ public class HeroTakeDamage : TakeDamageController
         HealController.Instance.HPBar(hsm);
         hsm.heroPanelHandler.UpdateHeroPanel();
     }
-    private void DamagePop(HeroStateMachine hsm, float getDamageAmount)
+    private void DamagePop(HeroStateMachine hsm, int getDamageAmount)
     {
         bool isCritical = Random.Range(0, 100) < 5;
         if (isCritical) getDamageAmount *= 2;
         DamagePopup.Create(this.transform.Find("Body").position, getDamageAmount, isCritical, false);
         hsm.baseHero.curHP -= getDamageAmount;
     }
-    public override void TakeDamage(GameObject target, float getDamageAmount, BaseAttack.Effect attackType1, BaseAttack.Effect attackType2)
+    public override void TakeDamage(GameObject target, int getDamageAmount, BaseAttack.Effect attackType1, BaseAttack.Effect attackType2)
     {
         throw new System.NotImplementedException();
     }
